Initial attempts which did not pass the test
The "Inefficient" file passes completeness test cases, but fails on efficiency/runtime requirements for large inputs.

The "Pairs Approach" was abandoned before completion. Demarcating the negative space was an interesting approach, but this implementation turned out cumbersome, and with some more time on the drawing board, I realized that a paired down version of the first attempt should be the ideal approach.